use --port/-p for publishing ports now instead of api-port
