+++ main.go
@@ -83,8 +83,9 @@ func main() {
 					Usage: "Mount one or more volumes into every node of the cluster (Docker notation: `source:destination`)",
 				},
 				cli.StringSliceFlag{
-					Name:  "publish, add-port",
-					Usage: "Publish k3s node ports to the host (Format: `[ip:][host-port:]container-port[/protocol]@node-specifier`, use multiple options to expose more ports)",
+					// TODO: remove publish/add-port soon, to clean up
+					Name:  "port, p, publish, add-port",
+					Usage: "Publish k3s node ports to the host (Format: `-p [ip:][host-port:]container-port[/protocol]@node-specifier`, use multiple options to expose more ports)",
 				},
 				cli.IntFlag{
 					Name:  "port-auto-offset",
@@ -92,10 +93,9 @@ func main() {
 					Usage: "Automatically add an offset (* worker number) to the chosen host port when using `--publish` to map the same container-port from multiple k3d workers to the host",
 				},
 				cli.StringFlag{
-					// TODO: only --api-port, -a soon since we want to use --port, -p for the --publish/--add-port functionality
-					Name:  "api-port, a, port, p",
+					Name:  "api-port, a",
 					Value: "6443",
-					Usage: "Specify the Kubernetes cluster API server port (Format: `[host:]port` (Note: --port/-p will be used for arbitrary port mapping as of v2.0.0, use --api-port/-a instead for setting the api port)",
+					Usage: "Specify the Kubernetes cluster API server port (Format: `-a [host:]port`",
 				},
 				cli.IntFlag{
 					Name:  "wait, t",
